requirements.txt
aiogram==3.13.1
python-docx==1.1.2
python-dotenv==1.0.1
aiosqlite==0.20.0
